pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "lucian-core"
version = "0.2.0"
description = "Self-evolving AI persona — dreams, reflects, sets directives, and archives growth in long-term memory."
requires-python = ">=3.11"
dependencies = [
    "openai>=1.13",
    "python-dotenv",
    "pyyaml",
    "chromadb",
]

[project.optional-dependencies]
server = ["fastapi", "uvicorn"]
dev = [
    "pytest>=8.0",
    "pytest-cov",
    "mypy",
    "types-PyYAML",
]

# ── pytest ──────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

# ── mypy ────────────────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "lucian.*"
disallow_untyped_defs = true

# ── setuptools ──────────────────────────────────────────────────────────────
[tool.setuptools.packages.find]
include = ["lucian*"]
